Web_App/requirements.txt
click>=7.1.2
Flask>=1.1.2
gunicorn>=20.0.4
itsdangerous>=1.1.0
Jinja2>=2.11.3
MarkupSafe>=1.1.1
Werkzeug>=1.0.1
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
seaborn>=0.11.0
pandas
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0